Removed `EitherExtensions.kt` and replaced all `getOrThrow` usages with `getOrElse { throw it }`. Updated package structure for modularization by moving `ehic`, `pid`, and `mdl` components into `attestation` package and updating all references accordingly.
