+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/IssuerContext.kt
@@ -35,15 +35,28 @@ import eu.europa.ec.eudi.pidissuer.adapter.input.web.MetaDataApi
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.WalletApi
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.security.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.IssuerSigningKey
-import eu.europa.ec.eudi.pidissuer.adapter.out.ehic.GetEuropeanHealthInsuranceCardDataMock
-import eu.europa.ec.eudi.pidissuer.adapter.out.ehic.IssueSdJwtVcEuropeanHealthInsuranceCard
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.ehic.GetEuropeanHealthInsuranceCardDataMock
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.ehic.IssueSdJwtVcEuropeanHealthInsuranceCard
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.learningcredential.IssueLearningCredential
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.*
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.DefaultEncodeMobileDrivingLicenceInCbor
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.EncodeMobileDrivingLicenceInCbor
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.GetMobileDrivingLicenceDataMock
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.IssueMobileDrivingLicence
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.*
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.AdministrationClient
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.Credentials
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.DefaultEncodePidInCbor
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.EncodePidInCbor
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.GetPidDataFromKeyCloak
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.IsoCountry
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.IssueMsoMdocPid
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.IssueSdJwtVcPid
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.Realm
 import eu.europa.ec.eudi.pidissuer.adapter.out.jose.*
-import eu.europa.ec.eudi.pidissuer.adapter.out.learningcredential.IssueLearningCredential
-import eu.europa.ec.eudi.pidissuer.adapter.out.mdl.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.nonce.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.persistence.InMemoryDeferredCredentialRepository
 import eu.europa.ec.eudi.pidissuer.adapter.out.persistence.R2dbcIssuedCredentialRepository
-import eu.europa.ec.eudi.pidissuer.adapter.out.pid.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.proof.ValidateAttestationProof
 import eu.europa.ec.eudi.pidissuer.adapter.out.proof.ValidateJwtProofWithKeyAttestation
 import eu.europa.ec.eudi.pidissuer.adapter.out.proof.VerifyKeyAttestation
@@ -124,7 +137,7 @@ import kotlin.time.Duration.Companion.minutes
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.toJavaDuration
 import kotlin.time.toKotlinDuration
-import eu.europa.ec.eudi.pidissuer.adapter.out.ehic.IssuingCountry as EhicIssuingCountry
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.ehic.IssuingCountry as EhicIssuingCountry
 import io.ktor.client.HttpClient as KtorHttpClient
 import io.ktor.client.engine.java.Java as JavaEngine
 import java.time.Duration as JavaDuration
