+++ src/test/kotlin/eu/europa/ec/eudi/pidissuer/adapter/input/web/WalletApiTest.kt
@@ -33,6 +33,19 @@ import com.nimbusds.jwt.SignedJWT
 import com.nimbusds.oauth2.sdk.token.DPoPAccessToken
 import eu.europa.ec.eudi.pidissuer.PidIssuerApplicationTest
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.security.DPoPTokenAuthentication
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.City
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.EncodePidInCbor
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.FamilyName
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.GivenName
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.IsoCountry
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.IssuingAuthority
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.Nationality
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.Pid
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.PidMetaData
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.PidMsoMdocScope
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.PidSdJwtVcScope
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.PlaceOfBirth
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.State
 import eu.europa.ec.eudi.pidissuer.adapter.out.pid.*
 import eu.europa.ec.eudi.pidissuer.domain.*
 import eu.europa.ec.eudi.pidissuer.jwtProof
